third_party/blink/web_tests/fast/dom/HTMLLinkElement/link-preload-unsupported-destination-expected.txt

CONSOLE WARNING: line 6: <link rel=preload> must have a valid `as` value
CONSOLE WARNING: line 7: <link rel=preload> uses an unsupported `as` value
CONSOLE WARNING: line 8: <link rel=preload> uses an unsupported `as` value
CONSOLE WARNING: line 9: <link rel=preload> uses an unsupported `as` value
